feat(character_gallery): 支持符卡和技能冷却/消耗为小数点
- 将 SpellCardTierItem 中的 cost 和 cooldown 从 int 改为 double 类型
- 更新 JSON 序列化逻辑，使用 toDouble() 替代 toInt()
- 添加 _formatNumber() 函数格式化数值显示，整数不显示小数点
- 更新所有符卡、技能、终极技能的冷却和消耗显示逻辑
- 优化搜索框提示文本，改为更通用的"搜索..."
- 在 character_skill_card.dart 中应用数值格式化函数

+++ lib/core/models/character_models.dart
@@ -1363,8 +1363,8 @@ class SpellCardTierItem extends Equatable {
   final SpellCardTier tier;
   final String? description;
   final String? iconUrl;
-  final int? cost;
-  final int? cooldown;
+  final double? cost;
+  final double? cooldown;
   final String? damage;
 
   const SpellCardTierItem({
